revert some commits and fix ItMiiSampleFmwAux failure
