+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItMiiSampleFmwMain.java
@@ -25,6 +25,8 @@ import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
 @IntegrationTest
 class ItMiiSampleFmwMain {
 
+  private static ItMiiSampleHelper myItMiiSampleHelper = null;
+
   /**
    * Install Operator.
    * @param namespaces list of namespaces created by the IntegrationTestWatcher by the
@@ -32,14 +34,15 @@ class ItMiiSampleFmwMain {
    */
   @BeforeAll
   public static void init(@Namespaces(4) List<String> namespaces) {
-    ItMiiSampleHelper.initAll(namespaces, ItMiiSampleHelper.DomainType.JRF, ItMiiSampleHelper.ImageType.MAIN);
+    myItMiiSampleHelper = new ItMiiSampleHelper();
+    myItMiiSampleHelper.initAll(namespaces, ItMiiSampleHelper.DomainType.JRF, ItMiiSampleHelper.ImageType.MAIN);
   }
 
   /**
    * Test to verify MII sample JRF initial use case.
-   * Deploys a database and initializes it for RCU, 
-   * uses an FMW infra base image instead of WLS 
-   * base image, and uses a WDT model that's 
+   * Deploys a database and initializes it for RCU,
+   * uses an FMW infra base image instead of WLS
+   * base image, and uses a WDT model that's
    * specialized for JRF, but is otherwise similar to
    * the WLS initial use case.
    * @see ItMiiSampleWlsMain#testWlsInitialUseCase for more...
@@ -49,7 +52,7 @@ class ItMiiSampleFmwMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_JRF_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample JRF initial use case")
   void testFmwInitialUseCase() {
-    ItMiiSampleHelper.callInitialUseCase();
+    myItMiiSampleHelper.callInitialUseCase(this.getClass().getSimpleName().toLowerCase());
   }
 
 
@@ -62,7 +65,8 @@ class ItMiiSampleFmwMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_JRF_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample JRF update1 use case")
   void testFmwUpdate1UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update1", "Update1 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update1", "Update1 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -74,7 +78,8 @@ class ItMiiSampleFmwMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_JRF_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample JRF update2 use case")
   void testFmwUpdate2UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update2", "Update2 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update2", "Update2 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -86,8 +91,8 @@ class ItMiiSampleFmwMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_JRF_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample JRF update3 use case")
   void testFmwUpdate3UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update3-image,-check-image-and-push,-update3-main",
-        "Update3 use case failed");
+    myItMiiSampleHelper.callUpdateUseCase("-update3-image,-check-image-and-push,-update3-main",
+        "Update3 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -102,7 +107,8 @@ class ItMiiSampleFmwMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_JRF_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample JRF update4 use case")
   void testFmwUpdate4UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update4", "Update4 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update4", "Update4 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -111,6 +117,6 @@ class ItMiiSampleFmwMain {
   @AfterAll
   public void tearDownAll() {
     // db cleanup or deletion and uninstall traefik
-    ItMiiSampleHelper.tearDownAll();
+    myItMiiSampleHelper.tearDownAll();
   }
 }
