+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItMiiSampleWlsMain.java
@@ -25,6 +25,8 @@ import org.junit.jupiter.api.condition.DisabledIfEnvironmentVariable;
 @IntegrationTest
 class ItMiiSampleWlsMain {
 
+  private static ItMiiSampleHelper myItMiiSampleHelper = null;
+
   /**
    * Install Operator.
    * @param namespaces list of namespaces created by the IntegrationTestWatcher by the
@@ -32,7 +34,8 @@ class ItMiiSampleWlsMain {
    */
   @BeforeAll
   public static void init(@Namespaces(3) List<String> namespaces) {
-    ItMiiSampleHelper.initAll(namespaces, ItMiiSampleHelper.DomainType.WLS, ItMiiSampleHelper.ImageType.MAIN);
+    myItMiiSampleHelper = new ItMiiSampleHelper();
+    myItMiiSampleHelper.initAll(namespaces, ItMiiSampleHelper.DomainType.WLS, ItMiiSampleHelper.ImageType.MAIN);
   }
 
   /**
@@ -44,15 +47,15 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_CHECK_SAMPLE", matches = "true")
   @DisplayName("Test to verify MII Sample source")
   void testCheckMiiSampleSource() {
-    ItMiiSampleHelper.callCheckMiiSampleSource("-check-sample",
+    myItMiiSampleHelper.callCheckMiiSampleSource("-check-sample",
         "Sample source doesn't match with the generated source");
   }
 
   /**
-   * Test to verify MII sample WLS initial use case. 
+   * Test to verify MII sample WLS initial use case.
    * Builds image required for the initial use case, creates secrets, and
    * creates domain resource.
-   * Verifies all WebLogic Server pods are ready, are at the expected 
+   * Verifies all WebLogic Server pods are ready, are at the expected
    * restartVersion, and have the expected image.
    * Verifies the sample application is running
    * (response includes "Hello World!").
@@ -62,14 +65,14 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_WLS_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample WLS initial use case")
   void testWlsInitialUseCase() {
-    ItMiiSampleHelper.callInitialUseCase();
+    myItMiiSampleHelper.callInitialUseCase(this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
-   * Test to verify WLS update1 use case. 
-   * Adds a data source to initial domain via a configmap and updates the 
+   * Test to verify WLS update1 use case.
+   * Adds a data source to initial domain via a configmap and updates the
    * domain resource restartVersion.
-   * Verifies all WebLogic Server pods roll to ready, roll to the expected 
+   * Verifies all WebLogic Server pods roll to ready, roll to the expected
    * restartVersion, and have the expected image.
    * Verifies the sample application is running
    * and detects the new datasource (response includes
@@ -80,7 +83,8 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_WLS_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample WLS update1 use case")
   void testWlsUpdate1UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update1", "Update1 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update1", "Update1 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -98,7 +102,8 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_WLS_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample WLS update2 use case")
   void testWlsUpdate2UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update2", "Update2 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update2", "Update2 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -106,7 +111,7 @@ class ItMiiSampleWlsMain {
    * Deploys an updated WebLogic application to the running
    * domain from update1 using an updated Docker image,
    * and updates the domain resource restartVersion.
-   * Verifies all WebLogic Server pods roll to ready, roll to the expected 
+   * Verifies all WebLogic Server pods roll to ready, roll to the expected
    * restartVersion, and have the expected image.
    * Verifies the sample application is running
    * and is at the new version (response includes "v2").
@@ -116,8 +121,8 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_WLS_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample WLS update3 use case")
   void testWlsUpdate3UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update3-image,-check-image-and-push,-update3-main",
-        "Update3 use case failed");
+    myItMiiSampleHelper.callUpdateUseCase("-update3-image,-check-image-and-push,-update3-main",
+        "Update3 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -132,7 +137,8 @@ class ItMiiSampleWlsMain {
   @DisabledIfEnvironmentVariable(named = "SKIP_WLS_SAMPLES", matches = "true")
   @DisplayName("Test to verify MII sample WLS update4 use case")
   void testWlsUpdate4UseCase() {
-    ItMiiSampleHelper.callUpdateUseCase("-update4", "Update4 use case failed");
+    myItMiiSampleHelper
+        .callUpdateUseCase("-update4", "Update4 use case failed", this.getClass().getSimpleName().toLowerCase());
   }
 
   /**
@@ -141,6 +147,6 @@ class ItMiiSampleWlsMain {
   @AfterAll
   public void tearDown() {
     // uninstall traefik
-    ItMiiSampleHelper.tearDownAll();
+    myItMiiSampleHelper.tearDownAll();
   }
 }
